pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixo"
version = "0.1.0"
description = "Ollama for Computer Vision — run any heavy CV model on any laptop without freezing"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["computer-vision", "yolo", "sam2", "cli", "inference", "gpu", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "huggingface-hub>=0.20.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
yolo = ["ultralytics>=8.0.0", "opencv-python>=4.8.0"]
onnx = ["onnxruntime>=1.16.0"]
cloud = ["kaggle>=1.6.0"]
vision = ["transformers>=4.40.0", "torch>=2.0", "opencv-python>=4.8.0", "matplotlib>=3.7"]
all = ["pixo[yolo,onnx,cloud,vision]"]

[project.urls]
Homepage = "https://github.com/pixo-cv/pixo"
Repository = "https://github.com/pixo-cv/pixo"
Issues = "https://github.com/pixo-cv/pixo/issues"

[project.scripts]
pixo = "pixo.cli:app"

[tool.setuptools.packages.find]
include = ["pixo*"]
